backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
langchain
langchain-groq
langchain-community
langchain-text-splitters
pinecone-client
sentence-transformers
PyPDF2
docx2txt
langchain-pinecone
